+++ lib/src/logger.dart
@@ -55,6 +55,9 @@ class Logger {
   /// root [Logger].
   StreamController<LogRecord>? _controller;
 
+  /// Controller used to notify when the log level of this logger is changed.
+  StreamController<Level?>? _levelChangedController;
+
   /// Create or find a Logger by name.
   ///
   /// Calling `Logger(name)` will return the same instance whenever it is called
@@ -78,8 +81,12 @@ class Logger {
     if (name.startsWith('.')) {
       throw ArgumentError("name shouldn't start with a '.'");
     }
+    if (name.endsWith('.')) {
+      throw ArgumentError("name shouldn't end with a '.'");
+    }
+
     // Split hierarchical names (separated with '.').
-    var dot = name.lastIndexOf('.');
+    final dot = name.lastIndexOf('.');
     Logger? parent;
     String thisName;
     if (dot == -1) {
@@ -135,7 +142,26 @@ class Logger {
       throw UnsupportedError(
           'Cannot set the level to `null` on a logger with no parent.');
     }
+    final isLevelChanged = _level != value;
     _level = value;
+    if (isLevelChanged) {
+      _levelChangedController?.add(value);
+    }
+  }
+
+  /// Returns a stream of level values set to this [Logger].
+  ///
+  /// You can listen for set levels using the standard stream APIs,
+  /// for instance:
+  ///
+  /// ```dart
+  /// logger.onLevelChanged.listen((level) { ... });
+  /// ```
+  /// A state error will be thrown if the level is changed
+  /// inside the callback.
+  Stream<Level?> get onLevelChanged {
+    _levelChangedController ??= StreamController<Level?>.broadcast(sync: true);
+    return _levelChangedController!.stream;
   }
 
   /// Returns a stream of messages added to this [Logger].
@@ -157,7 +183,7 @@ class Logger {
   }
 
   /// Whether a message for [value]'s level is loggable in this logger.
-  bool isLoggable(Level value) => (value >= level);
+  bool isLoggable(Level value) => value >= level;
 
   /// Adds a log record for a [message] at a particular [logLevel] if
   /// `isLoggable(logLevel)` is true.
@@ -204,7 +230,7 @@ class Logger {
       }
       zone ??= Zone.current;
 
-      var record = LogRecord(
+      final record = LogRecord(
           logLevel, msg, fullName, error, stackTrace, zone, object, metaData);
 
       if (parent == null) {